added tables and figures to output. added session info to html.
pdf and csv files with the final results based on simulated data are small enough such that they can be shared in the repository. The estimation output for fullsample and K12fold CV is too larger and needs to be reproduced on the local machine of the user.

test_target is a target that indicates a succesful test for Make and rstan on the local machine.

tables_and_figures.html includes information about the device and R packages used to last produce these results.